refactor(perf): consolidate native-host perf under tests/native_host/perf
Move shell benchmarks, pytest harness, and compare/report tooling into
tests/native_host/perf/ as a sibling of e2e/. Lift cluster.py and
conftest.py to tests/native_host/ for shared fixtures.

run_perf.sh emits PERF_METRICS_JSON; pytest uses --perf-json; PR-vs-nightly
comparison runs via run_perf_compare.sh and python -m perf.report. Update
workflows and building.rst docs; remove perf_tests/ and e2e/perf_harness/.
