+++ .github/workflows/ci.yml
@@ -207,7 +207,6 @@ jobs:
           mkdir -p "$root"
           cp -r tests "$root/"
           cp -r scripts "$root/"
-          cp -r perf_tests "$root/"
 
       - name: Upload E2E assets
         uses: actions/upload-artifact@v7
