+++ .github/workflows/perf.yml
@@ -26,7 +26,7 @@ on:
         default: false
         type: boolean
       compare_threshold_pct:
-        description: Regression/improvement threshold percent (compare.py)
+        description: Regression/improvement threshold percent (perf.report)
         required: false
         default: "10"
         type: string
@@ -105,7 +105,6 @@ jobs:
           mkdir -p "$PERF_ASSETS_DIR"
           cp -r pr-src/tests "$PERF_ASSETS_DIR/"
           cp -r pr-src/scripts "$PERF_ASSETS_DIR/"
-          cp -r pr-src/perf_tests "$PERF_ASSETS_DIR/"
 
       - name: Set up Python venv
         run: |
@@ -144,58 +143,28 @@ jobs:
             " || true
           done
 
-      - name: Run perf (PR binaries)
+      - name: Run perf compare (PR vs nightly)
         env:
           SPUR_TEST_NODES: ${{ env.BM_NODES }}
           SPUR_TEST_SSH_USER: ${{ env.BM_SSH_USER }}
-          SPUR_TEST_BINARIES_DIR: /tmp/perf-pr-bins
-          SPUR_TEST_REMOTE_BIN_DIR: ${{ env.BM_REMOTE_BIN_DIR_PR }}
-          SPUR_PERF_SCRIPTS_DIR: ${{ env.PERF_ASSETS_DIR }}/perf_tests
-          SPUR_PERF_RUN_LABEL: pr-${{ inputs.pr_number }}
-          SPUR_PERF_JSON_OUT: /tmp/perf-pr.json
-        run: |
-          chmod +x /tmp/perf-pr-bins/*
-          pytest "$PERF_ASSETS_DIR/tests/native_host/e2e/test_scheduler_perf.py" -v -m perf -s
-
-      - name: Cleanup cluster between perf runs
-        run: |
-          IFS=',' read -ra NODES <<< "$BM_NODES"
-          for node in "${NODES[@]}"; do
-            ssh -o StrictHostKeyChecking=accept-new -o BatchMode=yes \
-              "${BM_SSH_USER}@${node}" "
-              sudo pkill -f '${BM_REMOTE_BIN_DIR_PR}/spurctld' 2>/dev/null || true
-              sudo pkill -f '${BM_REMOTE_BIN_DIR_PR}/spurd' 2>/dev/null || true
-              sudo pkill -f '${BM_REMOTE_BIN_DIR_NIGHTLY}/spurctld' 2>/dev/null || true
-              sudo pkill -f '${BM_REMOTE_BIN_DIR_NIGHTLY}/spurd' 2>/dev/null || true
-              sudo rm -rf /tmp/spur-e2e-* 2>/dev/null || true
-              rm -rf '${BM_REMOTE_BIN_DIR_PR}' '${BM_REMOTE_BIN_DIR_NIGHTLY}' 2>/dev/null || true
-            " || true
-          done
-
-      - name: Run perf (nightly binaries)
-        env:
-          SPUR_TEST_NODES: ${{ env.BM_NODES }}
-          SPUR_TEST_SSH_USER: ${{ env.BM_SSH_USER }}
-          SPUR_TEST_BINARIES_DIR: /tmp/perf-nightly-bins
-          SPUR_TEST_REMOTE_BIN_DIR: ${{ env.BM_REMOTE_BIN_DIR_NIGHTLY }}
-          SPUR_PERF_SCRIPTS_DIR: ${{ env.PERF_ASSETS_DIR }}/perf_tests
-          SPUR_PERF_RUN_LABEL: nightly
-          SPUR_PERF_JSON_OUT: /tmp/perf-nightly.json
-        run: |
-          chmod +x /tmp/perf-nightly-bins/*
-          pytest "$PERF_ASSETS_DIR/tests/native_host/e2e/test_scheduler_perf.py" -v -m perf -s
-
-      - name: Compare PR vs nightly
-        working-directory: ${{ env.PERF_ASSETS_DIR }}/tests/native_host/e2e
         run: |
           set -euo pipefail
-          echo "######## PERF COMPARISON (PR #${{ inputs.pr_number }} vs nightly) ########"
-          ARGS=(/tmp/perf-pr.json /tmp/perf-nightly.json)
-          ARGS+=(--threshold "${{ inputs.compare_threshold_pct }}")
+          chmod +x "$PERF_ASSETS_DIR/tests/native_host/perf/run_perf_compare.sh"
+          ARGS=(
+            /tmp/perf-pr-bins
+            /tmp/perf-nightly-bins
+            --candidate-remote-dir "${BM_REMOTE_BIN_DIR_PR}"
+            --baseline-remote-dir "${BM_REMOTE_BIN_DIR_NIGHTLY}"
+            --candidate-label "pr-${{ inputs.pr_number }}"
+            --baseline-label nightly
+            --candidate-json /tmp/perf-pr.json
+            --baseline-json /tmp/perf-nightly.json
+            --threshold "${{ inputs.compare_threshold_pct }}"
+          )
           if [ "${{ inputs.fail_on_regression }}" = "true" ]; then
             ARGS+=(--fail-on-regression)
           fi
-          python -m perf_harness.compare "${ARGS[@]}"
+          "$PERF_ASSETS_DIR/tests/native_host/perf/run_perf_compare.sh" "${ARGS[@]}"
 
       - name: Cleanup remote processes
         if: always()
