+++ tests/README.md
@@ -5,5 +5,5 @@ Test suites organized by deployment target. See [docs/developer/building.rst](..
 | Path | Description |
 |------|-------------|
 | `native_host/e2e/` | Deploys Spur on bare-metal nodes via SSH |
-| `native_host/e2e/perf_harness/` | Helpers for optional `@pytest.mark.perf` scheduler perf tests |
+| `native_host/perf/` | Optional `@pytest.mark.perf` scheduler perf tests and compare tooling |
 | `k8s/e2e/` | Deploys Spur into a Kubernetes cluster via SpurJob CRDs |
