+++ tests/native_host/perf/__init__.py
@@ -1,15 +1,16 @@
 # Copyright (c) 2026 Advanced Micro Devices, Inc. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-"""Native-host perf harness (optional ``@pytest.mark.perf`` tests)."""
+"""Native-host scheduler perf tests (optional ``@pytest.mark.perf``)."""
 
-from .compare import compare_perf_suites, format_comparison_report
-from .harness import (
+from perf.harness import parse_tiers_from_env, run_native_perf_suite
+from perf.report import (
     PerfSuiteResult,
     PerfTierResult,
+    compare_perf_suites,
+    format_comparison_report,
     format_perf_summary_report,
-    parse_tiers_from_env,
-    run_native_perf_suite,
+    load_suite_json,
     suite_to_dict,
     write_suite_json,
 )
@@ -20,6 +21,7 @@ __all__ = [
     "compare_perf_suites",
     "format_comparison_report",
     "format_perf_summary_report",
+    "load_suite_json",
     "parse_tiers_from_env",
     "run_native_perf_suite",
     "suite_to_dict",
